requirements.txt
starlette==0.12.0
pydantic==0.30
python-dotenv==0.10.3
fastapi==0.33.0
pyjwt
passlib[bcrypt]
py_eureka_client==0.7.2
requests==2.22.0
PyYAML==5.2
werkzeug==1.0.0
